requirements.txt
httplib2
bs4
python-dateutil